Rename banner file for consistency.
