+++ assets/js/components/Banner/index.stories.js
@@ -32,7 +32,7 @@ import enhancedMeasurementDesktopSVG from '@/svg/graphics/banner-enhanced-measur
 import enhancedMeasurementMobileSVG from '@/svg/graphics/banner-enhanced-measurement-setup-cta-mobile.svg?url';
 import signInWithGoogleDesktopSVG from '@/svg/graphics/banner-sign-in-with-google-setup-cta.svg?url';
 import signInWithGoogleMobileSVG from '@/svg/graphics/banner-sign-in-with-google-setup-cta-mobile.svg?url';
-import warningDesktopSVG from '@/svg/graphics/banner-warning.svg?url';
+import warningDesktopSVG from '@/svg/graphics/warning-banner.svg?url';
 import gatheringDataSVG from '@/svg/graphics/gathering-data.svg?url';
 import Link from '../Link';
