+++ assets/js/components/notifications/ZeroDataNotification.js
@@ -30,7 +30,7 @@ import { DAY_IN_SECONDS } from '../../util';
 import BannerNotification, {
 	TYPES,
 } from '@/js/googlesitekit/notifications/components/layout/BannerNotification';
-import warningDesktopSVG from '@/svg/graphics/banner-warning.svg?url';
+import warningDesktopSVG from '@/svg/graphics/warning-banner.svg?url';
 import { CORE_NOTIFICATIONS } from '@/js/googlesitekit/notifications/datastore/constants';
 
 export default function ZeroDataNotification( { id, Notification } ) {
